pyproject.toml
[project]
name = "ielts-analyzer-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "fastapi[all]>=0.115.12",
    "pydantic-ai-slim[groq]>=0.1.3",
]
